requirements.txt
streamlit
pandas
yfinance
supabase-py==1.0.3